# openapi.yaml
openapi: 3.1.0
info:
  title: Synoptic Data Generation API
  version: 1.0.0
  description: API for managing synthetic data generation projects and jobs.
servers:
  - url: /api/v1 # Base path for API endpoints relative to the backend server URL
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A short error code or key.
        message:
          type: string
          description: A human-readable error message.
        details:
          type: object
          additionalProperties: true
          description: Optional additional error details.
      required:
        - error
        - message
...
    ProjectSettings:
      type: object
      properties:
        dataRetentionDays:
          type: integer
          format: int32
          description: Number of days to retain generated data.
          default: 30
        maxStorageGB:
          type: integer
          format: int32
          description: Maximum storage quota in GB.
          default: 50
      required:
        - dataRetentionDays
        - maxStorageGB
    ProjectStorage:
      type: object
      properties:
        bucketName:
          type: string
          description: Name of the dedicated GCP Cloud Storage bucket.
        region:
          type: string
          description: GCP region where the bucket is located.
        usedStorageBytes:
          type: integer
          format: int64
          description: Current storage usage in bytes (updated periodically).
          readOnly: true # Indicate this is likely calculated/reported, not set by client
      required:
        - bucketName
        - region
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the project.
          readOnly: true
        name:
          type: string
          description: Name of the project.
        description:
          type: string
          description: Description of the project.
        customerId:
          type: string
          format: uuid # Assuming customers also have UUIDs
          description: ID of the customer who owns the project.
          readOnly: true # Usually set based on authenticated user during creation
        status:
          type: string
          enum: [active, archived] # Soft delete via status change? Or dedicated DELETE endpoint? Let's use status for now.
          description: Status of the project.
          default: active
        storage:
          $ref: '#/components/schemas/ProjectStorage'
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        teamMembers:
          type: object
          description: Map of User IDs to their roles within the project.
          additionalProperties:
            $ref: '#/components/schemas/Role'
          example:
            "user-id-1": "owner"
            "user-id-2": "admin"
            "user-id-3": "member"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the project was created.
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the project was last updated.
          readOnly: true
      required:
        - id
        - name
        - description
        - customerId
        - status
        - storage
        - settings
        - teamMembers
        - createdAt
        - updatedAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        # region: # Maybe region for bucket creation? To be decided in implementation.
        #  type: string
      required:
        - name
        - description
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        status:
          type: string
          enum: [active, archived]
          description: Update project status (e.g., for archiving).
      # Allow partial updates, so no required fields here
...
    Role:
      type: string
      enum: [owner, admin, member, viewer]
      description: Role of a user within a project.
    InviteMemberRequest:
      type: object
      properties:
        userId:
          type: string
          description: The ID of the user to invite.
        role:
          type: string
          # Allow inviting only as these roles
          enum: [admin, member, viewer]
          description: The role to assign to the invited user.
      required:
        - userId
        - role
...
  securitySchemes:
    BearerAuth: # Or could be cookie-based session
      type: http
      scheme: bearer
      bearerFormat: JWT # Or describe session cookie if used
    SessionCookieAuth: # Changed to reflect actual implementation
      type: apiKey
      in: cookie
      name: session_token # Name of the session cookie used by auth service
  responses: # Defined common responses here
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: [] # Apply security globally, can override per-operation
  - SessionCookieAuth: [] # Apply security globally using session cookies
paths:
...
  /projects:
    get:
      summary: List projects for the current user
      tags:
        - Projects
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived]
          description: Filter projects by status.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of projects to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of projects to skip for pagination.
      responses:
        '200':
          description: A list of projects.
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    description: Total number of projects matching the filter.
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new project
      tags:
        - Projects
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID of the project to operate on.
    get:
      summary: Get project details by ID
      tags:
        - Projects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Project details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User does not have access to this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch: # Or PUT if full replacement is required
      summary: Update project details
      tags:
        - Projects
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User does not have permission to update this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a project # Or maybe just "Archive" if using status?
      tags:
        - Projects
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Project deleted successfully. No content.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User does not have permission to delete this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /projects/{projectId}/team:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      summary: Invite a user to the project team
      tags:
        - Team Management
      security:
        - BearerAuth: []
      description: Adds an existing registered user to the project with a specified role. Requires Admin or Owner role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '200':
          description: Member invited successfully. Returns the updated project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...